backend/requirements.txt
fastapi
uvicorn[standard]
websockets
openai
groq
google-generativeai
python-dotenv
rich
colorama
pydantic
